requirements.txt
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
idna==2.10
lxml
pycodestyle==2.7.0
requests==2.25.1
toml==0.10.2
urllib3==1.26.6
pytest
Flask
python-decouple
flask-restful